test: add check for running process before killing and print logs if not running

+++ tests/post-test.sh
@@ -7,7 +7,12 @@ echo "================================================"
 echo "==> Shutting down server"
 BEDROCK_PID=$(<mofka.pid)
 
-kill $BEDROCK_PID
+if kill -0 $BEDROCK_PID 2>/dev/null; then
+    kill $BEDROCK_PID
+else
+    echo "Mofka isn't running anymore, here are the logs:"
+    cat mofka.err mofka.out
+fi
 
 sleep 1
 rm mofka.json mofka.pid
